Move str::escape_* to libcore
